+++ src/component/Contact/contact.css
@@ -1,0 +1,74 @@
+#contact {
+    background-color: #2a2c4c;
+    color: white;
+    height: 40vh;
+  }
+  
+  #contact h1{
+    padding-top: 10px;
+  }
+  
+  #contact p{
+    margin-top: 20px;
+  }
+  
+  .Singlecall {
+    max-width: 750px;
+  
+  }
+  
+  /* .social-media-icons-white-a {
+    color: white;
+    font-size: 20rem;
+    
+  } */
+  
+  .social-media-icons{
+    font-size: 30px;
+    /* display: flex; */
+    flex-direction: column;
+    margin-left: 38px;  
+    margin-top: -18%;
+    padding-top: 160px;
+    
+    
+    justify-content: space-evenly; 
+  
+  }
+  
+  /* .social-media-icons p{
+    color: white;
+    display: flex;
+    font-size: 10px;
+  } */
+  .icon{
+    color: white;
+    margin-right: 10px;
+    padding-top: 20px;
+  }
+  .icontwo{
+    color: white;
+    margin-right: 10px;
+  
+  }
+  .iconthree{
+    color: white;
+    margin-right: 10px;
+  
+  }
+  .iconfour{
+    color: white;
+    margin-right: 10px;
+  
+  }
+  .iconfive{
+    color: white;
+    margin-right: 10px;
+  
+  
+  }
+  
+  .footer{
+    height: 12vh;
+  }
+
