+++ src/component/Header/Header.jsx
@@ -1,5 +1,6 @@
 import React from "react";
-import Navbar from "./Navbar";
+import Navbar from "../Navbar/Navbar";
+
 
 export default function Header() {
   return (
