+++ src/component/Howitworks/Howitworks.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import "./howitworks.css"
 
 function Howitworks() {
   return (
