+++ src/component/Navbar/Navbar.jsx
@@ -1,31 +1,31 @@
 import React from "react";
 import { Link } from "react-scroll";
-import stawibrood from "../images/Stawibrood.png";
+// import {HashLink as Link} from 'react-router-hash-link';
+import "./navbar.css"
 
-import "./style.css";
 
 function Navbar() {
   return (
     <div className="navbar">
       <nav className="header">
         <div className="logo-nav">
-          <Link to="main" className="logo">
+          {/* <Link to="main" className="logo">
             <img src={stawibrood} alt="logo" />
-          </Link>
+          </Link> */}
         </div>
 
         <ul>
           <li>
-            <a href="#contact">Contact Us</a>
+            <a href="contact">Contact Us</a>
           </li>
           <li>
-            <a href="#Team">Our Team</a>
+            <a href="Team">Our Team</a>
           </li>
           <li>
             <a href="#product">Our product</a>
           </li>
           <li>
-            <a href="Navbar">About</a>
+            <a href="#Navbar">About</a>
           </li>
           
           
@@ -35,7 +35,7 @@ function Navbar() {
       <div className="stawi">
         <h2>StawiBrood is here to</h2>
         <h1>
-          Give your chicks a <br></br>fighting chance
+          give your chicks a <br></br>fighting chance
         </h1>
       </div>
     </div>
