+++ src/component/Ourteam/Ourteam.css
@@ -1,0 +1,20 @@
+.team {
+    display: flex;
+    margin-left: 5%;
+  }
+  .koyo {
+    margin-left: 5%;
+  }
+  .virgy {
+    margin-left: 6%;
+  }
+  .mitchie {
+    margin-left: 8%;
+  }
+  .nancy {
+    margin-left: 5%;
+  }
+  .ourteam {
+    margin-top: -20%;
+    margin-left: 30%;
+  }
